Added spring security with encrypted password, which is save in DB
