+++ sandbox-web/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8090
+server.port=8085
 spring.application.name=sandbox
 server.servlet.context-path=/api
 server.error.include-message=always
